Move KLAP protocol implementation into its own file
